requirements_orc.txt
acres==0.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
astor==0.8.1
async-timeout==4.0.3
attrdict==2.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bce-python-sdk==0.9.46
bcrypt==4.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
camelot-py==1.0.9
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
chromadb==1.1.0
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.2
cryptography==46.0.1
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.1.4
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
durationpy==0.10
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.3.0
filelock==3.19.1
fire==0.7.1
fitz==0.0.1.dev2
Flask==3.1.2
flask-babel==4.0.0
flatbuffers==25.2.10
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
future==1.0.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==33.1.0
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.29
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.29
lazy_loader==0.4
lmdb==1.7.3
looseversion==1.3.0
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==1.24.4
oauthlib==3.3.1
onnxruntime==1.22.1
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opt-einsum==3.3.0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
paddleocr==2.7.0.3
paddlepaddle-gpu==2.6.0.post117
pandas==2.3.2
pathlib==1.0.1
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.3.0
pip-tools==7.5.0
posthog==5.4.0
premailer==3.10.0
propcache==0.3.2
protobuf==3.20.2
prov==2.1.1
psutil==7.1.0
puremagic==1.30
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.20.2
pyparsing==3.2.5
pypdf==3.17.4
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
RapidFuzz==3.14.1
rarfile==4.2
rdflib==7.2.1
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
shapely==2.1.1
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
tifffile==2025.5.10
tokenizers==0.22.1
tomli==2.2.1
tqdm==4.67.1
traits==7.0.2
typer==0.19.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.36.0
visualdl==2.5.3
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
